+++ Elements/test/SolidTests.cs
@@ -22,6 +22,7 @@ namespace Elements.Tests
         public SolidTests(ITestOutputHelper output)
         {
             this.output = output;
+            this.GenerateIfc = false;
 
             if (!Directory.Exists("models"))
             {
@@ -396,6 +397,144 @@ namespace Elements.Tests
             var rotate = new Transform();
             rotate.Rotate(Vector3.XAxis, 15);
             b.Representation.SolidOperations[0].LocalTransform = rotate;
+        }
+
+        public void SolidIntersectsWithPlane()
+        {
+            this.Name = nameof(SolidIntersectsWithPlane);
+            var n = 4;
+            var outer = Polygon.Ngon(n, 2);
+            var inner = Polygon.Ngon(n, 1.75).Reversed();
+            var profile = new Profile(outer, new[] { inner });
+            var sweep = new Extrude(profile, 5, Vector3.ZAxis, false);
+
+            var plane1 = new Plane(new Vector3(0, 0, 1), new Vector3(0.5, 0.5, 0.5));
+            var plane2 = new Plane(new Vector3(0, 0, 2), new Vector3(0.1, 0, 1));
+            var plane3 = new Plane(new Vector3(0, 0, 5), Vector3.ZAxis);
+
+            Plane[] planes = new Plane[] { plane1, plane2, plane3 };
+            var r = new Random();
+            foreach (var plane in planes)
+            {
+                if (sweep.Solid.Intersects(plane, out List<Polygon> result))
+                {
+                    if (result.Count > 1)
+                    {
+                        Assert.Equal(2, result.Count);
+                        var cutProfile = new Profile(result[0], result.Skip(1).ToArray());
+                        var lam = new Lamina(cutProfile, false);
+                        var cutRep = new Representation(new List<SolidOperation>() { lam });
+                        this.Model.AddElement(new GeometricElement(representation: cutRep, material: r.NextMaterial()));
+                    }
+                    else
+                    {
+                        Assert.Single(result);
+                        this.Model.AddElement(new Panel(result[0], r.NextMaterial()));
+                    }
+                }
+            }
+
+            var rep = new Representation(new List<SolidOperation>() { sweep });
+            var solidElement = new GeometricElement(representation: rep, material: BuiltInMaterials.Mass);
+            this.Model.AddElement(solidElement);
+        }
+
+        [Fact]
+        public void SolidIntersectsPlaneAtFaceReturningFaceLoop()
+        {
+            this.Name = nameof(SolidIntersectsPlaneAtFaceReturningFaceLoop);
+
+            var r = new Random();
+            var profile = Polygon.Rectangle(5, 5);
+            var extrude = new Extrude(profile, 5, Vector3.ZAxis, false);
+            var plane = new Plane(new Vector3(0, -2.5, 2.5), Vector3.YAxis.Negate());
+            Assert.True(extrude.Solid.Intersects(plane, out List<Polygon> result));
+
+            Assert.Single(result);
+            var p = result[0];
+            Assert.Equal(4, p.Vertices.Count);
+
+            this.Model.AddElement(new Panel(p, r.NextMaterial()));
+
+            var rep = new Representation(new List<SolidOperation>() { extrude });
+            var solidElement = new GeometricElement(representation: rep, material: BuiltInMaterials.Mass);
+            this.Model.AddElement(solidElement);
+        }
+
+        [Fact]
+        public void SolidIntersectsPlaneAtVertexWithNoResult()
+        {
+            var profile = Polygon.Rectangle(5, 5);
+            var extrude = new Extrude(profile, 5, Vector3.ZAxis, false);
+            var plane = new Plane(new Vector3(2.5, 2.5, 5.0), new Vector3(0.5, 0.5, 0.5));
+            Assert.False(extrude.Solid.Intersects(plane, out List<Polygon> result));
+        }
+
+        [Fact]
+        public void SolidIntersectsPlaneAtEdgeWithNoResult()
+        {
+            var profile = Polygon.Rectangle(5, 5);
+            var extrude = new Extrude(profile, 5, Vector3.ZAxis, false);
+            var plane = new Plane(new Vector3(2.5, 2.5, 0.0), new Vector3(0.5, 0.5, 0.0));
+            Assert.False(extrude.Solid.Intersects(plane, out List<Polygon> result));
+        }
+
+        [Fact]
+        public void SolidIntersectPlaneTwice()
+        {
+            this.Name = nameof(SolidIntersectPlaneTwice);
+            var r = new Random();
+            var l = Polygon.L(5, 5, 1);
+            var profile = new Profile(l, l.Offset(-0.1).Reversed());
+
+            var arc = new Arc(Vector3.Origin, 5, 0, 180);
+            var sweep = new Sweep(profile, arc, 0, 0, 0, false);
+            var plane = new Plane(Vector3.Origin, Vector3.YAxis.Negate());
+            Assert.True(sweep.Solid.Intersects(plane, out List<Polygon> result));
+
+            Assert.Equal(4, result.Count);
+            foreach (var p in result)
+            {
+                Assert.Equal(6, p.Vertices.Count);
+                this.Model.AddElement(new Panel(p, r.NextMaterial()));
+            }
+
+            var rep = new Representation(new List<SolidOperation>() { sweep });
+            var solidElement = new GeometricElement(representation: rep, material: BuiltInMaterials.Mass);
+            this.Model.AddElement(solidElement);
+        }
+
+        [Theory]
+        [InlineData("SolidIntersectionTest1", "../../../models/Geometry/SolidPlaneIntersection/debug-case-1.json")]
+        [InlineData("SolidIntersectionTest2", "../../../models/Geometry/SolidPlaneIntersection/debug-case-2.json")]
+        [InlineData("SolidIntersectionTest3", "../../../models/Geometry/SolidPlaneIntersection/debug-case-3.json")]
+        public void SolidPlaneIntersectionTests(string name, string path)
+        {
+            this.Name = name;
+
+            var r = new Random();
+
+            var di = JsonConvert.DeserializeObject<DebugInfo>(File.ReadAllText(path), new[] { new SolidConverter() });
+            foreach (var solid in di.Solid)
+            {
+                Assert.True(di.Plane.Normal.IsUnitized());
+                Assert.True(solid.Intersects(di.Plane, out var results));
+
+                foreach (var p in results)
+                {
+                    this.Model.AddElement(new Panel(p, r.NextMaterial()));
+                }
+                var rep = new Representation(new List<SolidOperation>() { new ConstructedSolid(solid, false) });
+                var solidElement = new GeometricElement(representation: rep, material: BuiltInMaterials.Mass);
+                this.Model.AddElement(solidElement);
+            }
+        }
+
+        private class DebugInfo
+        {
+            public List<Solid> Solid { get; set; }
+            public Plane Plane { get; set; }
+            public string Exception { get; set; }
         }
     }