+++ CHANGELOG.md
@@ -20,10 +20,14 @@
 - `LocalClassification`
 - `ModelArrows`
 - `ModelText`
+- `Solid.Intersects(Plane p, out List<Polygon> result)`
+- `Vector3.IsUnitized()`
+- `Transform.Inverse()`
 
 ### Changed
 
 ### Fixed
+
 - Deduplicate catalog names during code generation.
 
 ## 0.9.2
@@ -58,7 +62,6 @@
   - `Line.ExtendTo(Polygon)`
   - `Line.ExtendTo(Profile)`
 - Support for DXF from many basic elements.
-
 
 ### Changed